+++ src/vue/common/icons/social/Facebook-icon.vue
@@ -1,0 +1,15 @@
+<template>
+  <!-- eslint-disable -->
+  <svg
+    viewBox="0 0 18 18"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M11.3395 3.99002H12.7495V1.60502C12.0668 1.53403 11.3809 1.49898 10.6945 1.50002C8.65453 1.50002 7.25953 2.74502 7.25953 5.02502V6.99002H4.95703V9.66002H7.25953V16.5H10.0195V9.66002H12.3145L12.6595 6.99002H10.0195V5.28752C10.0195 4.50002 10.2295 3.99002 11.3395 3.99002Z" fill="currentColor"/>
+  </svg>
+  <!-- eslint-enable -->
+</template>
+
+<script>
+export default { name: 'facebook-icon' }
+</script>